Use docker-java 0.10.3-SNAPSHOT
This version is needed because of bug fixes and improvements in
Bind.parse().

In addition, as DockerClientBuilder now (since 0.10.2) ignores empty
version strings, there is no more need to handle this case here.
Another change is that docker-java now uses ServiceLoader internally to
locate its implementation, which fails with the default classloader.